added a secure deleting system

+++ cms-pdo/classes/Article.php
@@ -103,18 +103,23 @@ class Article {
         $article = $this->getArticleById($id);
 
         if($article) {
-            if(!empty($article->image) && file_exists($article->image)) {
-                if(!unlink($article->image)) {
-                    return false;
+            if($article->user_id == $_SESSION['user_id']) {
+
+                if(!empty($article->image) && file_exists($article->image)) {
+                    if(!unlink($article->image)) {
+                        return false;
+                    }
                 }
+            
+                $query = "DELETE FROM " . $this->table . " WHERE id = :id";
+
+                $stmt = $this->conn->prepare($query);
+                $stmt->bindParam(':id', $id, PDO::PARAM_INT);
+
+                return $stmt->execute();
+            } else {
+                redirect("admin.php");
             }
-        
-        $query = "DELETE FROM " . $this->table . " WHERE id = :id";
-
-        $stmt = $this->conn->prepare($query);
-        $stmt->bindParam(':id', $id, PDO::PARAM_INT);
-
-        return $stmt->execute();
         }
 
         return false;
